backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.0
transformers==4.35.0
scikit-learn==1.3.2
pandas==2.1.2
numpy==1.26.1
python-docx
PyPDF2
beautifulsoup4
tqdm==4.66.1
requests==2.31.0
datasets==3.2.0
psutil==5.9.6
gputil==1.4.0
python-dotenv==1.0.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
